feat: redid server code and routes to make signup/login work
