+++ package.json
@@ -37,8 +37,10 @@
     "html-webpack-plugin": "2.29.0",
     "http": "0.0.0",
     "ipaddr.js": "^1.5.2",
+    "jade": "^1.11.0",
     "jest": "20.0.4",
     "jquery": "^3.2.1",
+    "morgan": "^1.9.0",
     "object-assign": "4.1.1",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
