+++ src/Api.js
@@ -4,7 +4,7 @@ var socket;
 instantiateSocket();
 
 function instantiateSocket() {
-    if (window.location.port == 3001) {
+    if (window.location.port === '3001') {
         socket = io(window.location.protocol + "//" + window.location.hostname + ":3000", { 'transports': ['websocket', 'polling'] });
     }
     else {
