+++ src/Users.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import $ from 'jquery';
 import './Users.css';
 
 class Users extends Component {
