+++ views/error.jade
@@ -1,0 +1,4 @@
+block content
+  h1= message
+  h2= error.status
+  pre #{error.stack}
